Fixes based on review
Made fixes based on review comments. The big change is that timer5 is slowed to a cadence of 5ms.

+++ src/modules/sensors.h
@@ -32,9 +32,9 @@ struct sensors {
 extern struct sensors sensors;
 
 // TODO: Doc
-int16_t sensorsModuleInit(void);
+int sensorsModuleInit(void);
 
 // TODO: Doc
-int16_t sensorsModuleRun(void);
+int sensorsModuleRun(void);
 
 #endif /* __SENSORS_MODULE_H__ */
